Fix failing golangci-lint-check Set up a timeout for golangci-lint checks. Address the majority of lint issues by simplifying code, sorting import orders, removing unused variables, applying missing Go documentation, fixing variable names, correcting typos, and improving the usage of 'switch' statements, among other things. Additionally, introduce a list of lint issues for the community to handle.

+++ pkg/api/server/db/pagination/pagination.go
@@ -90,7 +90,7 @@ func (b *Batcher) Update(matched, total int) {
 
 // Next returns the recommended next batch size to query.
 func (b *Batcher) Next() int {
-	n := int(math.Ceil(float64(b.want) / float64(b.ratio)))
+	n := int(math.Ceil(float64(b.want) / b.ratio))
 	if n > b.max {
 		return b.max
 	}
